@@ -18,7 +18,7 @@
     "lint:ci": "eslint --ext .ts src/app",
     "lite": "test -d ./dist/statistics-dashboard && lite-server --baseDir=./dist/statistics-dashboard || echo '(run dist first)'",
     "prep-test-data": "node tools/test-data-prep/rename-files-to-mjs.mjs ./out-tsc && (node tools/test-data-prep/remove-localisation.mjs ./out-tsc)",
-    "prettier": "prettier --write '{src,test-data}/**/*.{ts,js,json,scss,html}' '*.{js,json}'",
+    "prettier": "prettier --write '{src,cypress,test-data}/**/*.{ts,js,json,scss,html}' '*.{js,json}'",
     "prettier:ci": "prettier --list-different '{src,test-data}/**/*.{ts,js,json,scss,html}' '*.{js,json}'",
     "start:ci": "ng serve --port 4280 --live-reload=false --configuration=e2e",
     "start:ci-data": "tsc --project tsconfig.test-data.json && npm run prep-test-data && (node --experimental-specifier-resolution=node --use_strict ./out-tsc/app/test-data/index-x.mjs)",
@@ -32,37 +32,39 @@
   "dependencies": {
     "@amcharts/amcharts4": "^4.10.22",
     "@amcharts/amcharts4-geodata": "^4.1.21",
-    "@angular/animations": "^16.2.6",
+    "@angular/animations": "^17.1.1",
     "@angular/cdk": "^16.2.5",
-    "@angular/common": "^16.2.6",
-    "@angular/compiler": "^16.2.6",
-    "@angular/core": "^16.2.6",
-    "@angular/forms": "^16.2.6",
-    "@angular/localize": "^16.2.6",
+    "@angular/common": "^17.1.1",
+    "@angular/compiler": "^17.1.1",
+    "@angular/core": "^17.1.1",
+    "@angular/forms": "^17.1.1",
+    "@angular/localize": "^17.1.1",
     "@angular/material": "^16.2.5",
-    "@angular/platform-browser": "^16.2.6",
-    "@angular/platform-browser-dynamic": "^16.2.6",
-    "@angular/router": "^16.2.6",
-    "@europeana/metis-ui-consent-management": "^0.0.13",
-    "@europeana/metis-ui-maintenance-utils": "0.0.5",
-    "ngx-matomo": "^2.0.0",
+    "@angular/platform-browser": "^17.1.1",
+    "@angular/platform-browser-dynamic": "^17.1.1",
+    "@angular/router": "^17.1.1",
+    "@europeana/metis-ui-consent-management": "^0.0.16",
+    "@europeana/metis-ui-maintenance-utils": "^0.0.12",
+    "cypress-axe": "^1.5.0",
+    "ngx-cookie-service": "^17.0.1",
+    "ngx-matomo-client": "^6.0.2",
     "pdfmake": "^0.1.72",
     "rxjs": "^7.8.1",
     "ts-loader": "^9.3.1",
     "tslib": "^2.0.0",
-    "zone.js": "^0.13.3"
+    "zone.js": "^0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.2.3",
-    "@angular/cli": "^16.2.3",
-    "@angular/compiler-cli": "^16.2.6",
+    "@angular-devkit/build-angular": "^17.1.1",
+    "@angular/cli": "^17.1.1",
+    "@angular/compiler-cli": "^17.1.1",
     "@cypress/webpack-preprocessor": "^5.12.2",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "^2.0.10",
     "@types/node": "^12.20.26",
-    "@typescript-eslint/eslint-plugin": "^5.36.0",
+    "@typescript-eslint/eslint-plugin": "^6.20.0",
     "@typescript-eslint/eslint-plugin-tslint": "^5.36.0",
-    "@typescript-eslint/parser": "^5.36.0",
+    "@typescript-eslint/parser": "^6.20.0",
     "@typescript-eslint/types": "^5.36.0",
     "axe-core": "^4.4.1",
     "codelyzer": "^6.0.0",
@@ -81,7 +83,7 @@
     "npm-run-all": "^4.1.5",
     "prettier": "^2.4.1",
     "ts-node": "~8.3.0",
-    "typescript": "^5.1.6",
+    "typescript": "^5.3.3",
     "typescript-eslint": "0.0.1-alpha.0"
   }
 }
