@@ -1,7 +1,6 @@
 import { DimensionName } from '../../src/app/_models';
 
 context('Statistics Dashboard', () => {
-
   const ctZeroParam = 'content-tier-zero=true';
   const ctZeroFilterParam = DimensionName.contentTier;
   const selCtrlCTZero = '[for=ctZero]';
@@ -14,17 +13,24 @@ context('Statistics Dashboard', () => {
   const selTotalRecords = '[data-e2e="total-records"]';
 
   describe('content-tier zero', () => {
-
     it('should enable and disable the control according to the url', () => {
       cy.visit(urlDefault);
 
-      cy.get(selCtrlCTZero).contains('Disable content tier 0').should('have.length', 0);
-      cy.get(selCtrlCTZero).contains('Enable content tier 0').should('have.length', 1);
+      cy.get(selCtrlCTZero)
+        .contains('Disable content tier 0')
+        .should('have.length', 0);
+      cy.get(selCtrlCTZero)
+        .contains('Enable content tier 0')
+        .should('have.length', 1);
 
       cy.visit(urlCTZero);
 
-      cy.get(selCtrlCTZero).contains('Disable content tier 0').should('have.length', 1);
-      cy.get(selCtrlCTZero).contains('Enable content tier 0').should('have.length', 0);
+      cy.get(selCtrlCTZero)
+        .contains('Disable content tier 0')
+        .should('have.length', 1);
+      cy.get(selCtrlCTZero)
+        .contains('Enable content tier 0')
+        .should('have.length', 0);
     });
 
     it('should enable and disable the menu item', () => {
@@ -33,12 +39,12 @@ context('Statistics Dashboard', () => {
       cy.get(selFilterLabel).should('have.length', 0);
       cy.get(selFilterOpener).contains('Content Tier').should('have.length', 1);
 
-      cy.get(selFilterOpener).contains('Content Tier').click({force: true});
+      cy.get(selFilterOpener).contains('Content Tier').click({ force: true });
       cy.get(selFilterLabel).contains('Tier 0').should('have.length', 0);
 
-      cy.get(selCtrlCTZero).click({force: true});
+      cy.get(selCtrlCTZero).click({ force: true });
 
-      cy.get(selFilterOpener).contains('Content Tier').click({force: true});
+      cy.get(selFilterOpener).contains('Content Tier').click({ force: true });
       cy.get(selFilterLabel).contains('Tier 0').should('have.length', 1);
     });
 
@@ -48,7 +54,7 @@ context('Statistics Dashboard', () => {
       cy.get(selTotalRecords).contains('1,000').should('have.length', 1);
       cy.get(selTotalRecords).contains('819').should('have.length', 0);
 
-      cy.get(selCtrlCTZero).click({force: true});
+      cy.get(selCtrlCTZero).click({ force: true });
 
       cy.get(selTotalRecords).contains('1,000').should('have.length', 0);
       cy.get(selTotalRecords).contains('819').should('have.length', 1);
