@@ -1,8 +1,7 @@
 import { DimensionName } from '../../src/app/_models';
 
 context('Statistics Dashboard', () => {
-
-  const force = { force : true };
+  const force = { force: true };
   const selExports = '.export.active';
   const selOpener = '.export-opener';
   const selText = '.copy-source';
@@ -11,7 +10,6 @@ context('Statistics Dashboard', () => {
   const urlFiltered = `${urlDefault}?provider=Athena`;
 
   describe('exports', () => {
-
     it('should show and hide the exports panel', () => {
       cy.visit(urlDefault);
       cy.get(selExports).should('have.length', 0);
@@ -31,6 +29,5 @@ context('Statistics Dashboard', () => {
       cy.get(selOpener).click(force);
       cy.get(selText).invoke('val').should('contain', urlFiltered);
     });
-
   });
 });
