@@ -12,7 +12,7 @@
 // This function is called when a project is opened or re-opened (e.g. due to
 // the project's config changing)
 
-const webpack = require('@cypress/webpack-preprocessor');
+const webpack = require("@cypress/webpack-preprocessor");
 
 /**
  * @type {Cypress.PluginConfig}
@@ -23,8 +23,8 @@ module.exports = (on, config) => {
   // `config` is the resolved Cypress config
 
   const options = {
-    webpackOptions: require('../../tools/cypress-webpack.config.js'),
+    webpackOptions: require("../../tools/cypress-webpack.config.js"),
     watchOptions: {}
   };
-  on('file:preprocessor', webpack(options));
-}
+  on("file:preprocessor", webpack(options));
+};
