@@ -15,7 +15,7 @@
 
 import '@angular/localize/init';
 // Import commands.js using ES2015 syntax:
-import './commands'
+import './commands';
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
