@@ -6,6 +6,7 @@ import { ClickService } from '../../_services';
 import { ClickAwareDirective } from '.';
 
 @Component({
+  imports: [ClickAwareDirective],
   template: `
     <div class="cmp">
       <div class="dead-zone">
@@ -24,7 +25,8 @@ import { ClickAwareDirective } from '.';
       </div>
     </div>
   `,
-  styles: ['.collapsed{ background-color: red; }']
+  styles: ['.collapsed{ background-color: red; }'],
+  standalone: true
 })
 class TestClickAwareDirectiveComponent {
   @ViewChild('clickInfo') clickInfo: ClickAwareDirective;
@@ -43,7 +45,7 @@ describe('ClickAwareDirective', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ClickAwareDirective, TestClickAwareDirectiveComponent],
+      imports: [ClickAwareDirective, TestClickAwareDirectiveComponent],
       schemas: [CUSTOM_ELEMENTS_SCHEMA]
     }).compileComponents();
   }));
