@@ -1,11 +1,19 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
-import { FormControl, FormGroup } from '@angular/forms';
+import {
+  FormControl,
+  FormGroup,
+  FormsModule,
+  ReactiveFormsModule
+} from '@angular/forms';
 import { externalLinks } from '../_data';
+import { NgClass, NgIf } from '@angular/common';
 
 @Component({
   selector: 'app-ct-zero-control',
   templateUrl: './ct-zero-control.component.html',
-  styleUrls: ['./ct-zero-control.component.scss']
+  styleUrls: ['./ct-zero-control.component.scss'],
+  standalone: true,
+  imports: [FormsModule, ReactiveFormsModule, NgClass, NgIf]
 })
 export class CTZeroControlComponent {
   @Input() form: FormGroup<{ contentTierZero: FormControl<boolean> }>;
@@ -20,7 +28,6 @@ export class CTZeroControlComponent {
    * emits event (invokes updatePageUrl())
    **/
   valueChanged(): void {
-    console.log('the value changes here...............');
     this.onChange.emit();
   }
 }
