@@ -1,11 +1,14 @@
 import { Component, EventEmitter, Output } from '@angular/core';
 import { externalLinks } from '../_data/static-data';
 import { environment } from '../../environments/environment';
+import { RouterLink } from '@angular/router';
 
 @Component({
   selector: 'app-footer',
   templateUrl: './footer.component.html',
-  styleUrls: ['./footer.component.scss']
+  styleUrls: ['./footer.component.scss'],
+  standalone: true,
+  imports: [RouterLink]
 })
 export class FooterComponent {
   public externalLinks = externalLinks;
